scripts/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ecosystem-scripts"
dynamic = ["version"]
description = "scripts for ecosystem package data"
readme = "../README.md"
requires-python = ">=3.11"
license = "GPL-3.0"
authors = [{ name = "Philipp A.", email = "flying-sheep@web.de" }]
urls.Documentation = "https://github.com/scverse/ecosystem-packages#readme"
urls.Issues = "https://github.com/scverse/ecosystem-packages/issues"
urls.Source = "https://github.com/scverse/ecosystem-packages"
dependencies = [
    "rich",
    "PyGitHub",
    "PyYAML",
    "httpx",
    "pillow>=12.0.0",
    "jsonschema>=4.25.1",
]

[project.scripts]
register-template-repos = "ecosystem_scripts.template_repo_registry:main"
validate-registry = "ecosystem_scripts.validate_registry:main"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0"

[tool.hatch.envs.default]
python = "3.13"

[tool.ruff]
line-length = 120
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = []
